adds a .scss for likebutton and definePlugin webpack
Previously there was a problem if we were to test a component that contains a require to a .scss file: jest would go crazy with the different require and fail the tests. With a plugin set for defining the process.env variable we are able to tackle this issue.
This commit also adds a .scss for a component so that we are able to fully validate this.